8c: some important long -> int32
test plan:
cd tests/c/misc/
~/goken/compilers/8c/o.out -t -S minus_one_index.c
...
	TEXT	foo+0(SB),0,$4
	MOVL	s+-4(SP),AX
	MOVB	$0,-1(AX)
	RET	,
	END	,

+++ compilers/8c/txt.c
@@ -1372,9 +1372,11 @@ gpseudo(int a, Sym *s, Node *n)
 int
 sconst(Node *n)
 {
-	long v;
+	int32 v;
 
 	if(n->op == OCONST && !typefd[n->type->etype]) {
+        // can be converted from big number to negative one
+        // as downcast vlong to int32
 		v = n->vconst;
 		if(v >= -32766L && v < 32766L)
 			return 1;
